Show the mobile navigation on mobile
When viewing the NativePHP website on a mobile device, it is now difficult for users to navigate to the mobile page to purchase a licence.

This PR ensures the mobile navigation item is present on mobile devices - as currently the user meeds to navigate to the docs, then go to installation page and then select to obtain a licence.